requirements.txt
anthropic>=0.40.0
streamlit>=1.32.0
pymupdf>=1.24.0
python-docx>=1.1.0
reportlab>=4.1.0
pydantic>=2.6.0
pytest>=8.0.0
